scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
